[feat] 상세 목표 삭제 confirm dialog 추가
삭제할건지 한번 더 확인하는 dialog 추가 구현
