+++ sidebar/src/components/UI/molecules/confirm_dialog/style.js
