+++ sidebar/src/components/UI/molecules/detail_list_item/index.js
@@ -14,6 +14,7 @@ import IconButton from '@material-ui/core/IconButton';
 import CreateIcon from '@material-ui/icons/Create';
 import RemoveIcon from '@material-ui/icons/Remove';
 import CancelSaveButton from '../cancle_save_button';
+import ConfirmDialog from '../confirm_dialog';
 import { deleteDetail, updateDetailInfo } from '../../../../modules/details';
 import getFormatData from '../../../../lib/date';
 import { useStyles, DetailTextWrapper, InputWrapper } from './style';
@@ -21,15 +22,14 @@ import { useStyles, DetailTextWrapper, InputWrapper } from './style';
 const DetailListItem = ({ detail, handleToggle, checked }) => {
   const classes = useStyles();
   const [edit, setEdit] = useState(false);
+  const [open, setOpen] = useState(false);
   const [selectedDate, setSelectedDate] = useState(detail.dueDate);
   const [title, setTitle] = useState(detail.title);
   const [prevTitle, setPrevTitle] = useState('');
   const [prevSelectedDate, setPrevSelectedDate] = useState('');
   const dispatch = useDispatch();
 
-  const handleRemove = () => {
-    dispatch(deleteDetail({ no: detail.no }));
-  };
+  const handleRemove = () => dispatch(deleteDetail({ no: detail.no }));
 
   const handleModify = (e) => {
     e.stopPropagation(); // 이벤트 버블링 제거
@@ -38,13 +38,8 @@ const DetailListItem = ({ detail, handleToggle, checked }) => {
     setEdit(true);
   };
 
-  const handleDateChange = (date) => {
-    setSelectedDate(date);
-  };
-
-  const handleChange = ({ target }) => {
-    setTitle(target.value);
-  };
+  const handleDateChange = (date) => setSelectedDate(date);
+  const handleChange = ({ target }) => setTitle(target.value);
 
   const handleCancel = () => {
     setTitle(prevTitle);
@@ -57,6 +52,9 @@ const DetailListItem = ({ detail, handleToggle, checked }) => {
     setSelectedDate(getFormatData(selectedDate));
     setEdit(false);
   };
+
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
 
   return (
     <>
@@ -109,12 +107,18 @@ const DetailListItem = ({ detail, handleToggle, checked }) => {
             <IconButton edge="end" aria-label="modify" onClick={handleModify}>
               <CreateIcon />
             </IconButton>
-            <IconButton edge="end" aria-label="remove" onClick={handleRemove}>
+            <IconButton edge="end" aria-label="remove" onClick={handleOpen}>
               <RemoveIcon />
             </IconButton>
           </ListItemSecondaryAction>
         </ListItem>
       )}
+      <ConfirmDialog
+        open={open}
+        handleClose={handleClose}
+        handleClick={handleRemove}
+        text="정말 삭제하시겠습니까?"
+      />
     </>
   );
 };
